+++ lib/rblade/compiler/compiles_components.rb
@@ -73,7 +73,7 @@ module RBlade
 
     def compile_component_end component
       code = "_slot=RBlade::SlotManager.new(_out);_out=_c#{component[:index]}_swap;"
-      code << "_out<<#{ComponentStore.component(component[:name])}(_slot,_c#{component[:index]}_attr);"
+      code << "_out<<#{ComponentStore.component(component[:name])}(_slot,_c#{component[:index]}_attr,params,session,flash,cookies);"
       code << "_slot=nil;_c#{component[:index]}_swap=nil;_c#{component[:index]}_attr=nil;"
 
       code