+++ test/test_case.rb
@@ -15,7 +15,10 @@ class TestCase < Minitest::Test
     if expected_result
       foo = "FOO" # standard:disable Lint/UselessAssignment
       bar = "BAR" # standard:disable Lint/UselessAssignment
-      params = {email: "user@example.com"}
+      params = {email: "user@example.com"} # standard:disable Lint/UselessAssignment
+      session = {user_id: 4} # standard:disable Lint/UselessAssignment
+      flash = {notice: "Request successful"} # standard:disable Lint/UselessAssignment
+      cookies = {accept_cookies: true} # standard:disable Lint/UselessAssignment
       result = eval RBlade::RailsTemplate.new.call(nil, template) # standard:disable Security/Eval
 
       assert_equal expected_result, result